Fix spock_apply.c for PG14 and PG15.
This had a regression for PG 14.x, x > 0.

For PG15, scanint8() has moved and has a different interface --
consistent with the other widely used strtointNN() functions.[1]

We only have a total of three call-sites for scanint8() in the entire
codebase and the other two will require error handling control so there
it does not seem judicious to spend time creating compatibility
functions avoiding the need for the #ifdef directives around the
call-sites.